+++ src/main/java/com/example/prision/modules/service/ICategoryService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ICategoryService {
     public CategoryResponseDto insertCategory(CategoryRequestDto categoryRequestDto);
-    public CategoryResponseDto findCategoryById(int id);
+    public CategoryResponseDto findCategoryById(Long id);
     public List<CategoryResponseDto> findAllCategories();
     public CategoryResponseDto updateCategory(Long id, CategoryRequestDto categoryRequestDto);
     public void deleteCategoryById(Long id);
